Started working on retrieving Guest Wallet
