+++ Sources/AppCoinsSDK/Models/Local/Wallet/WalletLocalService.swift
@@ -9,10 +9,10 @@ import Foundation
 
 internal protocol WalletLocalService {
     
-    func getActiveWallet() -> Wallet?
-    func getWalletList() -> [Wallet]
-    func createNewWallet() throws -> Wallet?
-    func importWallet(keystore: String, password: String, privateKey: String, completion: @escaping (Result<Wallet?, WalletLocalErrors>) -> Void)
+    func getActiveWallet() -> ClientWallet?
+    func getWalletList() -> [ClientWallet]
+    func createNewWallet() throws -> ClientWallet?
+    func importWallet(keystore: String, password: String, privateKey: String, completion: @escaping (Result<ClientWallet?, WalletLocalErrors>) -> Void)
     func getPrivateKey(address: String) -> Data?
     func getWalletSyncingStatus() -> WalletSyncingStatus
     func updateWalletSyncingStatus(status: WalletSyncingStatus)
