+++ Sources/AppCoinsSDK/Models/Services/APPC/APPCServiceClient.swift
@@ -1,22 +1,23 @@
 //
-//  AppCoinTransactionClient.swift
-//  
+//  APPCServiceClient.swift
 //
-//  Created by aptoide on 16/05/2023.
+//
+//  Created by aptoide on 12/07/2024.
 //
 
 import Foundation
 
-internal class AppCoinTransactionClient : AppCoinTransactionService {
-    
+internal class APPCServiceClient : APPCService {
+
     private let endpoint: String
     
-    internal init(endpoint: String = BuildConfiguration.transactionServiceURL) {
+    internal init(endpoint: String = BuildConfiguration.APPCServiceURL) {
         self.endpoint = endpoint
     }
     
-    internal func getBalance(wa: String, result: @escaping (Result<AppCoinBalanceRaw, AppcTransactionError>) -> Void) {
-        if let url = URL(string: endpoint + "/wallet/\(wa)/info") {
+    internal func getGuestWallet(guestUID: String, result: @escaping (Result<GuestWalletRaw, APPCServiceError>) -> Void) {
+        let route = "/guest_wallet"
+        if let url = URL(string: endpoint + route + "?id=\(guestUID)") {
             let task = URLSession.shared.dataTask(with: url) { data, response, error in
                 if let error = error {
                     if let nsError = error as NSError?, nsError.code == NSURLErrorNotConnectedToInternet {
@@ -25,9 +26,8 @@ internal class AppCoinTransactionClient : AppCoinTransactionService {
                         result(.failure(.failed))
                     }
                 } else {
-                    if let data = data, let balance = try? JSONDecoder().decode(AppCoinBalanceRaw.self, from: data) {
-                        print(balance)
-                        result(.success(balance))
+                    if let data = data, let findResult = try? JSONDecoder().decode(GuestWalletRaw.self, from: data) {
+                        result(.success(findResult))
                     } else { result(.failure(.failed)) }
                 }
