+++ Sources/AppCoinsSDK/Models/Services/APPC/APPCServiceErrors.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by aptoide on 12/07/2024.
+//
+
+import Foundation
+
+internal enum APPCServiceError: Error {
+    case failed
+    case noInternet
+}
